[REF] point_of_sale, pos_*: move login screen to point_of_sale
*: pos_hr, pos_event, pos_loyalty, pos_online_payment, pos_restaurant,
pos_sale, pos_sale_loyalty

Before, login screen was available only when the pos_hr module was
installed. This commit moves the login screen to the point_of_sale
module, so it is available by default.

closes odoo/odoo#178996

Taskid: 4161131
Related: odoo/enterprise#69401

+++ addons/point_of_sale/models/pos_session.py
@@ -186,8 +186,8 @@ class PosSession(models.Model):
 
     def delete_opening_control_session(self):
         self.ensure_one()
-        if self.state != 'opening_control':
-            raise UserError(_("You can only cancel a session that is in opening control state."))
+        if self.state != 'opening_control' or len(self.order_ids) > 0:
+            raise UserError(_("You can only cancel a session that is in opening control state and has no orders."))
         self.sudo().unlink()
         return {
             'status': 'success',
